+++ tests/test_init.py
@@ -5,4 +5,4 @@ from faddr import __version__
 
 def test_version():
     """Test app version."""
-    assert __version__ == "0.2.3"
+    assert __version__ == "0.3.0"