+++ CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove duplicated ip addresses with parser's postprocessing.
 
+### Removed
+
+- Remove windows platform from tests for now
+
 ## [0.2.3] - 2022-04-05
 
 ### Fixed